Safari-611.1.10/LayoutTests/media/controls/elementOrder-expected.txt

This tests for the content in the status display.

This test only runs in DRT!


EVENT: canplaythrough
PASS: We are using the apple idiom
PASS: Rewind Button is to the right of Play Button
PASS: Timeline Box is to the right of Rewind Button
PASS: Mute Box is to the right of Timeline Box
PASS: Fullscreen Button is to the right of Mute Box

Testing finished.
